requirements.txt
aiofiles~=24.1.0
fastapi~=0.115.8
pillow~=11.1.0
python-multipart~=0.0.20